requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
fastapi==0.122.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gevent==25.9.1
geventhttpclient==2.3.5
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
locust==2.42.5
locust-cloud==1.29.4
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.1.2
namex==0.1.0
narwhals==2.12.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
protobuf==4.25.8
psutil==7.1.3
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.15.0
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.4
rich==14.2.0
rpds-py==0.29.0
scipy==1.15.3
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
streamlit==1.51.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.2.0
tf_keras==2.16.0
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websocket-client==1.9.0
Werkzeug==3.1.3
wrapt==2.0.1
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1